# simpler, faster reactor